refactor(aiAgent): ChatSessionPane 会话列表逻辑下沉至 HistoryChat，提取 SessionSearchCommand 独立组件
