+++ app/renderer/src/main/src/pages/ai-agent/ChatSessionPane/ChatSessionPane.module.scss
@@ -4,104 +4,6 @@
   min-width: 0;
   min-height: 0;
   overflow: hidden;
-  padding: 12px 0 0;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-
-  .chat-session-list-header {
-    flex-shrink: 0;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 12px 8px;
-
-    .header-title {
-      display: flex;
-      align-items: center;
-      gap: 4px;
-      min-width: 0;
-      color: var(--Colors-Use-Neutral-Text-3-Secondary);
-      font-size: 12px;
-      font-weight: 500;
-
-      .title-text {
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        min-width: 0;
-      }
-
-      .session-count-tag {
-        min-width: 24px;
-        justify-content: center;
-        font-variant-numeric: tabular-nums;
-      }
-
-      .source-filter {
-        display: flex;
-        align-items: center;
-        gap: 2px;
-        margin-left: 2px;
-        padding-left: 4px;
-        border-left: 1px solid var(--Colors-Use-Neutral-Border);
-        flex: 0 0 auto;
-      }
-
-      .source-filter-item {
-        width: 24px;
-        height: 24px;
-        padding: 0;
-        border: 1px solid transparent;
-        border-radius: 4px;
-        background: transparent;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        color: var(--Colors-Use-Neutral-Text-3-Secondary);
-        cursor: pointer;
-
-        svg {
-          width: 16px;
-          height: 16px;
-        }
-
-        &:hover {
-          background: var(--Colors-Use-Neutral-Bg-Hover);
-        }
-      }
-
-      .source-filter-item-local.source-filter-item-active {
-        color: var(--Colors-Use-Main-Primary);
-        background: var(--Colors-Use-Main-Bg);
-        border-color: var(--Colors-Use-Main-Border);
-      }
-
-      .source-filter-item-feishu.source-filter-item-active {
-        background: var(--Colors-Use-Purple-Bg);
-        border-color: var(--Colors-Use-Purple-Border);
-      }
-
-      .source-filter-item-dingtalk.source-filter-item-active {
-        background: var(--Colors-Use-Blue-Bg);
-        border-color: var(--Colors-Use-Blue-Border);
-      }
-    }
-
-    .header-actions {
-      display: flex;
-      align-items: center;
-      gap: 4px;
-    }
-  }
-
-  .chat-session-list {
-    flex: 1;
-    min-height: 0;
-    overflow: hidden;
-    display: flex;
-    flex-direction: column;
-  }
 }
 
 .session-search-modal {
@@ -111,84 +13,5 @@
       overflow: hidden;
       box-shadow: 0 16px 48px var(--Colors-Use-Basic-Shadow);
     }
-  }
-}
-
-.session-command {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  padding: 8px;
-  gap: 8px;
-
-  .command-input {
-    flex-shrink: 0;
-
-    .search-icon {
-      svg {
-        width: 16px;
-        height: 16px;
-        color: var(--Colors-Use-Neutral-Disable);
-      }
-    }
-
-    :global {
-      .ant-input-affix-wrapper,
-      .ant-input-affix-wrapper:hover,
-      .ant-input-affix-wrapper-focused {
-        padding: 0 12px;
-        border-radius: 4px;
-        border-color: transparent !important;
-        box-shadow: none !important;
-        background: var(--Colors-Use-Neutral-Bg) !important;
-
-        input {
-          background-color: var(--Colors-Use-Neutral-Bg) !important;
-        }
-      }
-    }
-  }
-
-  .command-list {
-    max-height: 360px;
-    overflow-y: auto;
-    padding: 0;
-  }
-
-  .command-empty {
-    padding: 24px 8px;
-    text-align: center;
-    color: var(--Colors-Use-Neutral-Text-3-Secondary);
-    font-size: 12px;
-    line-height: 16px;
-  }
-
-  .command-item {
-    height: 32px;
-    padding: 0 12px;
-    display: flex;
-    align-items: center;
-    border-radius: 8px;
-    cursor: pointer;
-    color: var(--Colors-Use-Neutral-Text-2-Primary);
-    font-size: 12px;
-    line-height: 16px;
-
-    .command-item-title {
-      flex: 1;
-      min-width: 0;
-    }
-  }
-
-  .command-item-active {
-    background: var(--Colors-Use-Neutral-Bg);
-  }
-
-  .command-item-current {
-    color: var(--Colors-Use-Main-Primary);
-  }
-
-  .command-item-current.command-item-active {
-    background: var(--Colors-Use-Neutral-Bg);
   }
 }
