+++ app/renderer/src/main/src/pages/ai-agent/ChatSessionPane/SessionSearchCommand/SessionSearchCommand.module.scss
@@ -1,0 +1,78 @@
+.session-command {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  padding: 8px;
+  gap: 8px;
+
+  .command-input {
+    flex-shrink: 0;
+
+    .search-icon {
+      svg {
+        width: 16px;
+        height: 16px;
+        color: var(--Colors-Use-Neutral-Disable);
+      }
+    }
+
+    :global {
+      .ant-input-affix-wrapper,
+      .ant-input-affix-wrapper:hover,
+      .ant-input-affix-wrapper-focused {
+        padding: 0 12px;
+        border-radius: 4px;
+        border-color: transparent !important;
+        box-shadow: none !important;
+        background: var(--Colors-Use-Neutral-Bg) !important;
+
+        input {
+          background-color: var(--Colors-Use-Neutral-Bg) !important;
+        }
+      }
+    }
+  }
+
+  .command-list {
+    max-height: 360px;
+    overflow-y: auto;
+    padding: 0;
+  }
+
+  .command-empty {
+    padding: 24px 8px;
+    text-align: center;
+    color: var(--Colors-Use-Neutral-Text-3-Secondary);
+    font-size: 12px;
+    line-height: 16px;
+  }
+
+  .command-item {
+    height: 32px;
+    padding: 0 12px;
+    display: flex;
+    align-items: center;
+    border-radius: 8px;
+    cursor: pointer;
+    color: var(--Colors-Use-Neutral-Text-2-Primary);
+    font-size: 12px;
+    line-height: 16px;
+
+    .command-item-title {
+      flex: 1;
+      min-width: 0;
+    }
+  }
+
+  .command-item-active {
+    background: var(--Colors-Use-Neutral-Bg);
+  }
+
+  .command-item-current {
+    color: var(--Colors-Use-Main-Primary);
+  }
+
+  .command-item-current.command-item-active {
+    background: var(--Colors-Use-Neutral-Bg);
+  }
+}
