+++ app/renderer/src/main/src/pages/ai-agent/historyChat/HistoryChat.module.scss
@@ -94,7 +94,7 @@
       .header-actions {
         display: flex;
         align-items: center;
-        gap: 8px;
+        gap: 4px;
       }
     }
